Moved Interfaces definition to a different file
